# Conflicts:
#	app_manager/appmanager/src/main/java/app/editors/manager/app/App.kt
#	app_manager/appmanager/src/main/java/app/editors/manager/di/component/AppComponent.kt
#	app_manager/appmanager/src/main/java/app/editors/manager/managers/utils/UiUtils.kt
#	app_manager/appmanager/src/main/java/app/editors/manager/ui/fragments/main/CloudsFragment.java
#	app_manager/appmanager/src/main/java/app/editors/manager/ui/views/custom/MainToolbar.kt
#	app_manager/appmanager/src/main/res/layout/fragment_choose_clouds.xml

+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/fragments/main/DocsBaseFragment.java
@@ -470,29 +470,31 @@ public abstract class DocsBaseFragment extends ListFragment implements DocsBaseV
     @Override
     public void onAcceptClick(CommonDialog.Dialogs dialogs, @Nullable String value, @Nullable String tag) {
         super.onAcceptClick(dialogs, value, tag);
-        if (tag != null) {
-            switch (tag) {
-                case DocsBasePresenter.TAG_DIALOG_BATCH_DELETE_CONTEXT:
-                    getPresenter().delete();
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_CONTEXT_RENAME:
-                    getPresenter().rename(value);
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_ACTION_SHEET:
-                    getPresenter().createDocs(value + "." + ApiContract.Extension.XLSX.toLowerCase());
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_ACTION_PRESENTATION:
-                    getPresenter().createDocs(value + "." + ApiContract.Extension.PPTX.toLowerCase());
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_ACTION_DOC:
-                    getPresenter().createDocs(value + "." + ApiContract.Extension.DOCX.toLowerCase());
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_ACTION_FOLDER:
-                    getPresenter().createFolder(value);
-                    break;
-                case DocsBasePresenter.TAG_DIALOG_BATCH_DELETE_SELECTED:
-                    getPresenter().deleteItems();
-                    break;
+        if (isResumed()) {
+            if (tag != null) {
+                switch (tag) {
+                    case DocsBasePresenter.TAG_DIALOG_BATCH_DELETE_CONTEXT:
+                        getPresenter().delete();
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_CONTEXT_RENAME:
+                        getPresenter().rename(value);
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_ACTION_SHEET:
+                        getPresenter().createDocs(value + "." + ApiContract.Extension.XLSX.toLowerCase());
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_ACTION_PRESENTATION:
+                        getPresenter().createDocs(value + "." + ApiContract.Extension.PPTX.toLowerCase());
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_ACTION_DOC:
+                        getPresenter().createDocs(value + "." + ApiContract.Extension.DOCX.toLowerCase());
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_ACTION_FOLDER:
+                        getPresenter().createFolder(value);
+                        break;
+                    case DocsBasePresenter.TAG_DIALOG_BATCH_DELETE_SELECTED:
+                        getPresenter().deleteItems();
+                        break;
+                }
             }
         }
     }
@@ -813,8 +815,8 @@ public abstract class DocsBaseFragment extends ListFragment implements DocsBaseV
     }
 
     @Override
-    public void onDownloadActivity() {
-        showDownloadFolderActivity();
+    public void onDownloadActivity(Uri uri) {
+        showDownloadFolderActivity(uri);
     }
 
     @Override
@@ -1230,7 +1232,7 @@ public abstract class DocsBaseFragment extends ListFragment implements DocsBaseV
                     break;
                 case PDF:
                     intent.setClassName(requireContext(), EditorsContract.PDF);
-                    startActivityForResult(intent, REQUEST_PDF);
+                    startActivity(intent);
                     break;
             }
         } catch (ActivityNotFoundException e) {
